requirements.txt
# Core
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

# Module A
ultralytics>=8.0.0
transformers>=4.36.0

# Module B
matplotlib>=3.7.0
scipy>=1.11.0

# Module C & D
openai-whisper>=20231117
sentence-transformers>=2.2.0
soundfile>=0.12.0

# Module E (PDF)
reportlab>=4.0.0

# CLI / config
PyYAML>=6.0.0

# Web UI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
